ADD product module, moved allocate domain tests here
